separate components for barcode scanning feature
