+++ src/routes/ScanBarcode/components/BarcodeInputForm.jsx
@@ -1,0 +1,22 @@
+function BarcodeInputForm({ value, handleOnchange, handleSubmit } ) {
+  return (
+    <>
+      <div>
+        <div>
+          <label style={{ color: '#000000' }}>Barcode</label>
+          <input
+            placeholder="Enter barcode number. Ex: 3017624010701"
+            value={value}
+            onChange={(e) => handleOnchange(e.target.value)}
+          />
+        </div>
+
+        <button className="upload-button w-100" onClick={handleSubmit}>
+          Get Barcode Information
+        </button>
+      </div>
+    </>
+  )
+}
+
+export default BarcodeInputForm;
